% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Horse_kicks.R
\docType{data}
\name{Kicks}
\alias{Kicks}
\title{Original published data on deaths from horse kicks in the Prussian army.}
\format{
There are two data frames, \code{Kicks} and \code{Corps_size}.

\code{Kicks} has one row for each year and corps.
\itemize{
\item \code{year} Day of the week
\item \code{corps} Day (number) in the month
\item \code{deaths} number of deaths in that corps in that year.
}

\code{Corps_size} has one row for each corps and gives the relative sizes of the various corps.
}
\source{
Transcribed from the original book.
}
\usage{
Kicks
}
\description{
These are famous data published by Ludwig von Bortekewitscch in his 1898 book,
\emph{Das Gesetz der Kleinen Zahlen} ("The law of small numbers"). 20 years of data are presented
for each of 14 army corps.
}
\details{
More data on the same corps are available in a more modern data set, with 35 years of data
in the \code{hkdeaths} data frame in the \code{{HorseKicks}} package. These revised data and some analysis
are described in Antony Unwin, Bill Venables (March 2025) "Horses for courses: A fresh look at the horse-kick data"
\emph{Significance} \strong{22}(2):30–35,
March 2025, \verb{https://doi.org/10.1093/jrssig/qmaf009}
}
\keyword{datasets}
