% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mortality_history.R
\docType{data}
\name{Mortality_history}
\alias{Mortality_history}
\title{Yearly mortality tables for the US from 1900 to 2010, by age and sex}
\format{
\itemize{
\item \code{year} the calendar year
\item \code{age} the age from 0 (newborns) to 119.
\item \code{mortality} the fraction of people of that age who died in that year
\item \code{sex}: M or F
}
}
\source{
US Social Security Administration as cited above
}
\usage{
Mortality_history
}
\description{
These are available through the US Social Security Adminstration
at \verb{//www.ssa.gov/oact/HistEst/PerLifeTables/2014/PerLifeTables2014.html}
}
\details{
For each calendar year (1900, 1901, ..., 2010) there are 240 rows, covering ages 0 to 119 for
males and females separately.
}
\keyword{datasets}
