% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mortality2021.R
\docType{data}
\name{SSA2021F}
\alias{SSA2021F}
\title{US Mortality tables for 2021}
\format{
The unit of observation is a year of age. The variables are
\itemize{
\item \code{age} the age from 0 (newborns) to 100.
\item \code{mortality} the fraction of people of that age who died in 2021
\item \code{lives} the number in a hypothetical population of 100,000 newborns as they progress through
the years of age.
\item \code{LE} the life expectancy (in years) at that age.
}
}
\source{
US Social Security Administration as cited above
}
\usage{
SSA2021F
}
\description{
The 2024 Trustees Reports from the US Social Security Administration
is available online. The \href{https://www.ssa.gov/oact/STATS/table4c6.html}{"life tables"} (from 2021)
give the mortality rate at each age of life for males and females separately.
}
\details{
The data frames \code{SSA2021F} and \code{SSA2021M} are mortality tables for
females and males separately.

The \href{https://www.ssa.gov/oact/HistEst/Population/2023/Population2023.html}{population tables}
cover many years, giving the total population at each age for each year, and even break them down by marital status.
For your convenience, I have packaged up the female and male life tables (\code{SSA2021F} and \code{SSA2021M})
and the multi-year population table (\code{SSA_population}).
}
\keyword{datasets}
