% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSA_population.R
\docType{data}
\name{SSA_population}
\alias{SSA_population}
\title{US yearly population data from 1941 to 2021 and then projected onward to 2100.}
\format{
The unit of observation is a year of age in a calendar year. The variables are
\itemize{
\item \code{year} the calendar year: 1941 through 2021 and projected forward to 2100.
\item \code{age} the age from 0 (newborns) to 100.
\item \code{total} the total US population of that age in that year
\item \code{f.tot} and \code{m.tot}: total population for males and females of that age in that year
\item \code{f.sin} and \code{m.sin}: populations of single persons
\item \code{f.mar} and \code{m.mar}: populations of married persons
\item \code{f.wid} and \code{m.wid}: populations of widowed persons
\item \code{f.div} and \code{m.div}: populations of divorced persons
}
}
\source{
US Social Security Administration as cited above
}
\usage{
SSA_population
}
\description{
The US Social Security Administration provides \href{https://www.ssa.gov/oact/HistEst/Population/2023/Population2023.html}{population tables}
cover many years, giving the total population at each age, breaking them down by sex and marital status.
}
\keyword{datasets}
