% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxi_trips.R
\docType{data}
\name{Taxi_trips}
\alias{Taxi_trips}
\title{Taxi trips in New York City}
\format{
Data frame with 9501 rows. Each row is one taxi trip in New York City
\itemize{
\item \code{distance} the trip distance in miles
\item \code{fare} the amount paid for the trip in US dollars. Unclear if this includes any tip.
\item \code{passengers} the number of passengers on the trip.
\item \code{pickup} the data/time of the start of the triop
\item \code{dropoff} the date/time of the end of the trip.-
}
}
\source{
Downloaded from the link above in 2015
}
\usage{
Taxi_trips
}
\description{
This is a subset of a much larger data set from the
New York City \href{http://www.nyc.gov/html/tlc/html/passenger/taxicab_rate.shtml}{Taxi and Limosine Commission}
about taxicab trips in New York City.
}
\keyword{datasets}
